+++ design-system/tokens/examples/constellation/token-card-brand.tsx
@@ -13,14 +13,6 @@ border: \`1px solid \${token('color.border.brand')}\`,
 hoverBackgroundColor: token('color.background.brand.bold.hovered'),
 activeBackgroundColor: token('color.background.brand.bold.pressed'),
 iconColor: token('color.icon.inverse'),
-
-// default styles
-color: token('color.text.brand'),
-backgroundColor: token('color.background.brand'),
-border: \`1px solid \${token('color.border.brand')}\`,
-hoverBackgroundColor: token('color.background.brand.hovered'),
-activeBackgroundColor: token('color.background.brand.pressed'),
-iconColor: token('color.icon.brand'),
 `;
 
 const brandStyles = {
@@ -37,14 +29,6 @@ const brandStyles = {
       '#09326C',
     ),
     iconColor: token('color.icon.inverse', '#FFFFFF'),
-  },
-  default: {
-    color: token('color.text.brand', '#0C66E4'),
-    backgroundColor: token('color.background.brand', '#E9F2FF'),
-    border: `1px solid ${token('color.border.brand', '#0C66E4')}`,
-    hoverBackgroundColor: token('color.background.brand.hovered', '#CCE0FF'),
-    activeBackgroundColor: token('color.background.brand.pressed', '#85B8FF'),
-    iconColor: token('color.icon.brand', '#0C66E4'),
   },
 };