+++ media/smart-card/src/__tests__/visual-regression/flexible-card.test.tsx
@@ -169,7 +169,7 @@ describe('Flexible Card', () => {
       const url = getURL('vr-flexible-ui-element-text-and-date');
       const page = await setup(url);
       await page.waitForSelector('[data-testid="vr-test-text"]');
-      const image = await takeSnapshot(page, 230);
+      const image = await takeSnapshot(page, 300);
 
       expect(image).toMatchProdImageSnapshot();
     });